+++ lib/react_on_rails/react_component/render_options.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "react_on_rails/utils"
-require "react_on_rails/pro_utils"
 
 module ReactOnRails
   module ReactComponent
@@ -15,13 +14,10 @@ module ReactOnRails
       # TODO: remove the required for named params
       def initialize(react_component_name: required("react_component_name"), options: required("options"))
         @react_component_name = react_component_name.camelize
-
-        result = ReactOnRails::ProUtils.disable_pro_render_options_if_not_licensed(options)
-        @options = result[:raw_options]
-        @explicitly_disabled_pro_options = result[:explicitly_disabled_pro_options]
+        @options = options
       end
 
-      attr_reader :react_component_name, :explicitly_disabled_pro_options
+      attr_reader :react_component_name
 
       def throw_js_errors
         options.fetch(:throw_js_errors, false)
@@ -100,7 +96,11 @@ module ReactOnRails
       end
 
       def immediate_hydration
-        retrieve_configuration_value_for(:immediate_hydration)
+        ReactOnRails::Utils.normalize_immediate_hydration(
+          options[:immediate_hydration],
+          react_component_name,
+          "Component"
+        )
       end
 
       def to_s